chore: moved test util out of __test__
